Added all adapters and started writing endpoints

+++ Domain/Repositories/IUnitOfWork.cs
@@ -1,5 +1,5 @@
 ﻿namespace Domain.Repositories;
 
 public interface IUnitOfWork {
-  Task<long> CompleteAsync();
+  Task<int> CompleteAsync();
 }
